+++ packages/bluesky/test/src/unspecced/unspecced_service_test.dart
@@ -14,6 +14,7 @@ import 'package:test/test.dart';
 import 'package:bluesky/src/entities/feed.dart';
 import 'package:bluesky/src/entities/feed_generators.dart';
 import 'package:bluesky/src/entities/skeleton_feed.dart';
+import 'package:bluesky/src/entities/skeleton_posts_by_query.dart';
 import 'package:bluesky/src/unspecced/unspecced_service.dart';
 
 void main() {
@@ -291,6 +292,96 @@ void main() {
 
       atp_test.expectRateLimitExceededException(
         () async => await feeds.findTimelineSkeleton(),
+      );
+    });
+  });
+
+  group('.searchPostsByQuerySkeleton', () {
+    test('normal case', () async {
+      final unspecced = UnspeccedService(
+        atproto: ATProto(did: 'test', accessJwt: 'test'),
+        did: '',
+        protocol: Protocol.https,
+        service: 'test',
+        context: ClientContext(
+          accessJwt: '1234',
+          timeout: Duration.zero,
+        ),
+        mockedGetClient: atp_test.createMockedGetClient(
+          'test/src/unspecced/data/search_posts_by_query_skeleton.json',
+        ),
+      );
+
+      final response = await unspecced.searchPostsByQuerySkeleton('test');
+
+      expect(response, isA<XRPCResponse>());
+      expect(response.data, isA<SkeletonPostsByQuery>());
+
+      expect(response.data.posts.length, 2);
+      expect(response.data.hitsTotal, 10);
+      expect(response.data.cursor, 'xxxxxxx');
+    });
+
+    test('as JSON', () async {
+      final unspecced = UnspeccedService(
+        atproto: ATProto(did: 'test', accessJwt: 'test'),
+        did: '',
+        protocol: Protocol.https,
+        service: 'test',
+        context: ClientContext(
+          accessJwt: '1234',
+          timeout: Duration.zero,
+        ),
+        mockedGetClient: atp_test.createMockedGetClient(
+          'test/src/unspecced/data/search_posts_by_query_skeleton.json',
+        ),
+      );
+
+      final response = await unspecced.searchPostsByQuerySkeletonAsJson('test');
+
+      expect(response, isA<XRPCResponse>());
+      expect(response.data, isA<Map<String, dynamic>>());
+    });
+
+    test('when unauthorized', () async {
+      final unspecced = UnspeccedService(
+        atproto: ATProto(did: 'test', accessJwt: 'test'),
+        did: '',
+        protocol: Protocol.https,
+        service: 'test',
+        context: ClientContext(
+          accessJwt: '1234',
+          timeout: Duration.zero,
+        ),
+        mockedGetClient: atp_test.createMockedGetClient(
+          'test/src/data/error.json',
+          statusCode: 401,
+        ),
+      );
+
+      atp_test.expectUnauthorizedException(
+        () async => await unspecced.searchPostsByQuerySkeleton('test'),
+      );
+    });
+
+    test('when rate limit exceeded', () async {
+      final unspecced = UnspeccedService(
+        atproto: ATProto(did: 'test', accessJwt: 'test'),
+        did: '',
+        protocol: Protocol.https,
+        service: 'test',
+        context: ClientContext(
+          accessJwt: '1234',
+          timeout: Duration.zero,
+        ),
+        mockedGetClient: atp_test.createMockedGetClient(
+          'test/src/data/error.json',
+          statusCode: 429,
+        ),
+      );
+
+      atp_test.expectRateLimitExceededException(
+        () async => await unspecced.searchPostsByQuerySkeleton('test'),
       );
     });
   });